vnc: change VncDisplay from unsigned to signed integer
I find it somewhat odd to have a 0 display for a VNC (which is
absolutely valid and the default case) as an indicator of `no VNC`.

This patch makes the variable signed, so -1 (the default setting) means
`no VNC`. All other positive values including 0 are valid.

+++ tests/vnc/vnc_test.go
@@ -222,7 +222,7 @@ func TestVNCVMStart(t *testing.T) {
 
 	opts = append(opts, expect.WithMetadata(*metadata))
 
-	opts = append(opts, expect.WithVnc(uint32(*vncDisplay)))
+	opts = append(opts, expect.WithVnc(*vncDisplay))
 
 	opts = append(opts, expect.WithVncPassword(*vncPassword))
